Authentication Added
I Have added a google firebase authentication system. In this system users can log in and their name and image will be shown in the header as a profile, besides this, it has a dynamic navbar where login time the sign-in and sign-up buttons will be invisible but after logout, these buttons will be visible again. If the user is registered by google then his or her official photo will be visible in the header section other wise their email and an avatar picture will be shown.